+++ packages/ai-core/src/common/variable-service.ts
@@ -122,16 +122,10 @@ export namespace AIVariableResolutionRequest {
 export interface AIVariableContext {
 }
 
-export interface AIVariableDropResult {
-    variables: AIVariableResolutionRequest[],
-    text?: string
-};
-
 export type AIVariableArg = string | { variable: string, arg?: string } | AIVariableResolutionRequest;
 
 export type AIVariableArgPicker = (context: AIVariableContext) => MaybePromise<string | undefined>;
 export type AIVariableArgCompletionProvider = (model: monaco.editor.ITextModel, position: monaco.Position) => MaybePromise<monaco.languages.CompletionItem[] | undefined>;
-export type AIVariableDropHandler = (event: DragEvent, context: AIVariableContext) => Promise<AIVariableDropResult | undefined>;
 
 export interface AIVariableResolver {
     canResolve(request: AIVariableResolutionRequest, context: AIVariableContext): MaybePromise<number>,
@@ -159,13 +153,10 @@ export interface AIVariableService {
     unregisterArgumentCompletionProvider(variable: AIVariable, argPicker: AIVariableArgCompletionProvider): void;
     getArgumentCompletionProvider(name: string): Promise<AIVariableArgCompletionProvider | undefined>;
 
-    registerDropHandler(handler: AIVariableDropHandler): Disposable;
-    unregisterDropHandler(handler: AIVariableDropHandler): void;
-    getDropResult(event: DragEvent, context: AIVariableContext): Promise<AIVariableDropResult>;
-
     resolveVariable(variable: AIVariableArg, context: AIVariableContext): Promise<ResolvedAIVariable | undefined>;
 }
 
+/** Contributions on the frontend can optionally implement `FrontendVariableContribution`. */
 export const AIVariableContribution = Symbol('AIVariableContribution');
 export interface AIVariableContribution {
     registerVariables(service: AIVariableService): void;
@@ -177,7 +168,6 @@ export class DefaultAIVariableService implements AIVariableService {
     protected resolvers = new Map<string, AIVariableResolver[]>();
     protected argPickers = new Map<string, AIVariableArgPicker>();
     protected argCompletionProviders = new Map<string, AIVariableArgCompletionProvider>();
-    protected dropHandlers = new Set<AIVariableDropHandler>();
 
     protected readonly onDidChangeVariablesEmitter = new Emitter<void>();
     readonly onDidChangeVariables: Event<void> = this.onDidChangeVariablesEmitter.event;
@@ -299,30 +289,6 @@ export class DefaultAIVariableService implements AIVariableService {
 
     async getArgumentCompletionProvider(name: string): Promise<AIVariableArgCompletionProvider | undefined> {
         return this.argCompletionProviders.get(this.getKey(name)) ?? undefined;
-    }
-
-    registerDropHandler(handler: AIVariableDropHandler): Disposable {
-        this.dropHandlers.add(handler);
-        return Disposable.create(() => this.unregisterDropHandler(handler));
-    }
-
-    unregisterDropHandler(handler: AIVariableDropHandler): void {
-        this.dropHandlers.delete(handler);
-    }
-
-    async getDropResult(event: DragEvent, context: AIVariableContext): Promise<AIVariableDropResult> {
-        let text: string | undefined = undefined;
-        const variables: AIVariableResolutionRequest[] = [];
-        for (const handler of this.dropHandlers) {
-            const result = await handler(event, context);
-            if (result) {
-                variables.push(...result.variables);
-                if (text === undefined) {
-                    text = result.text;
-                }
-            }
-        }
-        return { variables, text };
     }
 
     async resolveVariable(request: AIVariableArg, context: AIVariableContext): Promise<ResolvedAIVariable | undefined> {
